Speed up flush evaluation
We use a 32 bit split in 4 parts (5 bit  each) inited
with 11 and add 1 for every card to the slot
according to card's color, if one slot reaches 16, then
we have a flush.

It is faster becuase it avoids 4 popocount() for each
hand evaluation.

Speed up of additional 1.2% with PGO compile.