Speed up enumeration
Instead of permutations of cards use combinations
for each hole and common cards.

Given permutation and combination formulas:

P(n, k) = n(n-1) ... (n - k - 1)
C(n, k) = n(n-1) ... (n - k - 1) / k(k - 1)...1

For a single set of 2-hole cards we have
C(64, 2) = 2016 instead of P(64,2) = 4032 enumerations.

For the 5 common cards we have
C(64, 5) = 7.624.512 instead of P(64,5) = 914.941.440 enumerations,
so 120 times fewer samples.

So speed up depends on numer of missing cards and in which
groups (holes or commons), but it is important and allows
us to lift the enum limit from 4 cards to 5.

Still something is different from PokerStove...