Fix split results computation
Split results in wins and ties and fix the way
ties are  accounted.

Instead of split the point in 2, split it according
to the number of winners: 2-way splits = 1/2, 3-way = 1/3 etc

To avoid a floating point division in a hot path, pre-multiply
by constant 60, choose it because it has integer quotient when
divided by 2,3,4,5,6 so that tie value is correct when up
to 6 players split the pot.

Verified tie scores are now correct when compared with PokerStove.