Enable ranges
Now it is possible to run something like:

$ ./poker go -p 3 [AA,QQ-99,AKs,T7s-T3s,AKo] [88+,T6s+,52o+] TT+

Where ranges are defined like T6s+, 88+, 52o+, AA, AK, QQ-99,
T7s-T3s, J8-52, etc.

Adding ranges management introduces a slowdown of less than 2%,
both when ranges are used or not. Indeed when using a range
instead of a free set of hole cards we are even a bit faster.