Fix enumeration bug with single card
With single card, full enumeration is wrong:

$ ./poker_ms.exe go -g 5M -p 2 Jh - 4h 5s 6s Ts

     Equity    Win     Tie   Pots won  Pots tied
P1:  36.24%  35.43%   0.81%   1771598  40335.00
P2:  63.76%  62.95%   0.81%   3147732  40335.00

$ ./poker_ms.exe go -e -g 5M -p 2 Jh - 4h 5s 6s Ts
Evaluating 2140380 combinations...

     Equity    Win     Tie   Pots won  Pots tied
P1:  32.11%  31.52%   0.58%    674722  12453.00
P2:  67.89%  67.31%   0.58%   1440752  12453.00

This is because random numbers for hole cards are
generated prepending newest cards, while are picked
in reversed order (from LSB to MSB) during run().

Bug does not affect pair free-cards or ranges, it
appears only when there is a hole with a single
free card and one or more holes with double
free cards.